👽 api : detail api 수정
