+++ client/src/pages/Detail.js
@@ -1,9 +1,10 @@
-import React from 'react';
-// import { useEffect } from 'react'
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
-// import axios from 'axios';
+import axios from 'axios';
 // import { useRecoilValue } from 'recoil';
+import { useRecoilState, useRecoilValue } from 'recoil';
 
+import { DetailInformation, PlaceIDState } from '../atoms';
 import ReservationAsideBar from '../components/ReservationComponents/ReservationAsideBar';
 import Nav from '../components/Navigation/Nav';
 import View from '../components/View';
@@ -24,14 +25,41 @@ function Detail() {
   //     .catch(err => console.log(err));
   // }, []); // > 실제 api
 
+  const [detailInformation, setDetailInformation] =
+    useRecoilState(DetailInformation);
+  const placeId = useRecoilValue(PlaceIDState);
+  const header = {
+    headers: {
+      'ngrok-skip-browser-warning': '010',
+    },
+  };
+
+  const callDetailData = async () => {
+    try {
+      const response = await axios.get(`/place/${placeId}`, header);
+      setDetailInformation({ ...response.data });
+      // const { title, filePath, category, detailInfo, address, phoneNumber } =
+      //   detailInformation;
+      // console.log(response.data);
+      // return [title, filePath, category, detailInfo, address, phoneNumber];
+    } catch (error) {
+      // return console.log(error);
+      console.log(error);
+    }
+  };
+
+  useEffect(() => {
+    callDetailData();
+  }, []);
+
   return (
     <DetailReviewContainer>
       <DetailContainer>
         <DetailViewContainer>
           <Nav />
-          <View />
+          <View detailInformation={detailInformation} />
         </DetailViewContainer>
-        <ReservationAsideBar />
+        <ReservationAsideBar charge={detailInformation.charge} />
       </DetailContainer>
       <ReviewContainer />
     </DetailReviewContainer>
